+++ apps/remnant2toolkit/app/(items)/_constants/ring-items.ts
@@ -395,7 +395,7 @@ export const ringItems: RingItem[] = [
     id: 'w268ge',
     description: `Increases Movement Speed and Physical Damage by 5%. Bonus is doubled when HASTE is active.`,
     wikiLinks: [`https://remnant.wiki/Breach_Accelerator`],
-    location: { world: `N'Erud`, dungeon: 'World Drop' },
+    location: { world: `N'Erud`, dungeon: ['Agronomy Sector'] },
     dlc: 'dlc3',
   },
   {