+++ apps/remnant2toolkit/app/guides/faq/layout.tsx
@@ -1,0 +1,44 @@
+import { type Metadata } from 'next';
+
+import { OG_IMAGE_URL, SITE_TITLE } from '@/app/_constants/meta';
+import { NAV_ITEMS } from '@/app/_constants/nav-items';
+
+export async function generateMetadata(): Promise<Metadata> {
+  const title = `${NAV_ITEMS.faqGuide.label} - ${SITE_TITLE}`;
+  const description = NAV_ITEMS.faqGuide.description;
+
+  return {
+    title,
+    description,
+    openGraph: {
+      title,
+      description,
+      siteName: SITE_TITLE,
+      url: `https://remnant2toolkit.com/${NAV_ITEMS.faqGuide.href}`,
+      images: [
+        {
+          url: OG_IMAGE_URL,
+          width: 150,
+          height: 150,
+        },
+      ],
+      type: 'website',
+    },
+    twitter: {
+      title,
+      description,
+    },
+  };
+}
+
+export default async function Layout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="flex w-full max-w-2xl flex-col items-start justify-start text-left">
+      {children}
+    </div>
+  );
+}